Move page_slug + site_root into jottit/urls.py
Both are URL-shape helpers — neither belongs in render.py (where
page_slug was) or views/page.py (where _site_root lived as a private
helper). Co-locating them in a dedicated module also makes the shared
"site_root() + page_slug(name)" pattern easier to spot.
